+++ frontend/src/components/styles/Home.css
